backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.1.0
django-filter==25.2
